# Description
I have managed to improve the performance of the pool and improve its readability

# Benchmarks
## Old
```
BenchmarkEventPoolOneGoroutine-8        58460538                19.96 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8           100          10746244 ns/op            4986 B/op         56 allocs/op
BenchmarkEventPoolSlowestPath-8              152           8189235 ns/op           27087 B/op       1034 allocs/op
```
## New
```
BenchmarkEventPoolOneGoroutine-8        39543708                32.61 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8           697           1724738 ns/op             876 B/op         15 allocs/op
BenchmarkEventPoolSlowestPath-8             1026           1038654 ns/op           24619 B/op       1007 allocs/op
```