@ythosa 2024-04-22
btw, i've tested realisation based on `sync.Pool`, it was too much faster than my new realisation, but the possibility of event dump was lost