@vadimalekseev 2024-04-26
@ythosa, hello!

You can use BenchmarkLightJsonReadPar to check how the performance has changed. On mac m1 this bench showed following result:
old:
```
BenchmarkLightJsonReadPar-8            1        2256751708 ns/op         643.06 MB/s     4825008 B/op      48384 allocs/op
```
new:
```
BenchmarkLightJsonReadPar-8            1        4433031459 ns/op         327.37 MB/s     4272744 B/op      42659 allocs/op
```

Also, we can't use sync.Pool because insaneJSON uses own global pool. So, we need to refactor insaneJSON first, then we can test sync.Pool